UserAvatar.vue: add UserAvatar component
